Move txgh patch into ext directory
