+++ lib/txdb/tx_resource.rb
@@ -1,20 +1,3 @@
-module Txgh
-  class TxResource
-    class << self
-      def from_api_response(project_slug, response)
-        new(
-          project_slug,
-          response['slug'],
-          response['i18n_type'],
-          response['source_language_code'],
-          response['name'],
-          '', nil
-        )
-      end
-    end
-  end
-end
-
 module Txdb
   class TxResource
     attr_reader :original, :content
